backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
httpx>=0.27.0
pydantic>=2.0.0
PyJWT>=2.8.0
pyjwt[crypto]>=2.8.0
python-dotenv>=1.0.0
mcp>=0.1.0
playwright>=1.40.0
